Remove CAPE-specific things.
CAPE-specific modules have been moved to the CAPE repo and are no
longer needed here.
