+++ zerok/zerok_lib/src/wallet/mod.rs
@@ -1,5 +1,5 @@
-pub mod cape;
 pub mod network;
-pub mod spectrum;
 #[cfg(any(test, feature = "mocks"))]
 pub mod testing;
+
+pub use seahorse::*;
