+++ zerok/zerok_lib/src/wallet/testing/mod.rs
@@ -1,7 +1,4 @@
 #![allow(dead_code)]
 
-pub mod mocks;
 pub use seahorse::testing::*;
-
-mod cape_test;
-mod spectrum_test;
+pub mod mocks;
